requirements.txt
pynvml>=11.5.0
pyyaml>=6.0
numpy>=1.24
pandas>=2.0
pytest>=7.4
